删除属性时检查是否有商品使用 删除商品时删除商品分、商品相关、商品sku、商品属性、商品描述

+++ beike/Admin/Repositories/AttributeRepo.php
@@ -13,6 +13,7 @@ namespace Beike\Admin\Repositories;
 
 use Beike\Models\Attribute;
 use Beike\Models\AttributeValue;
+use Beike\Models\ProductAttribute;
 
 class AttributeRepo
 {
@@ -107,6 +108,10 @@ class AttributeRepo
 
     public static function delete($id)
     {
+        $productIds = ProductAttribute::query()->where('attribute_id', $id)->pluck('product_id')->toArray();
+        if ($productIds) {
+            throw New \Exception(trans('admin/attribute.error_cannot_delete_product_used', ['product_ids' => implode(', ', $productIds)]));
+        }
         $attribute = Attribute::query()->findOrFail($id);
         $attribute->descriptions()->delete();
         $attribute->values()->delete();
